This is Darren I added the map and I added in props
The props are in the Starter Content Folder and it's called Mesh. If you
use the meshes, theres not texture, the proper textures are inside the
mesh folder. I didn't want to put the textures into the material folder
becuase the it will be eaiser to find the correct texture if i was
inside the folder